safari-600.7-branch/LayoutTests/fast/canvas/canvas-clearRect-expected.txt

Series of tests to ensure correct behavior of canvas.clearRect().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test canvas.clearRect() with height = width = 0.
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
Test canvas.clearRect() with height = 0, width = 1.
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
Test canvas.clearRect() with height = 1, width = 0.
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS successfullyParsed is true

TEST COMPLETE
